moved test folder into ./src
