+++ server/package.json
@@ -2,7 +2,7 @@
   "name": "server",
   "version": "0.0.0",
   "private": true,
-  "type": "module",
+  "type": "commonjs",
   "scripts": {
     "build": "tsc --build",
     "clean": "tsc --build --clean",
